Repository: added getAllProductLines(), byDescription

+++ src/main/java/com/example/classicmodlesslaes/repository/ProductLineRepositoryImpl.java
@@ -3,10 +3,13 @@ package com.example.classicmodlesslaes.repository;
 import com.example.classicmodlesslaes.model.ProductLine;
 import com.example.classicmodlesslaes.repository.interfaces.ProductLineRepository;
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.Transient;
+import jakarta.persistence.TypedQuery;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Repository
 public class ProductLineRepositoryImpl implements ProductLineRepository {
@@ -19,27 +22,52 @@ public class ProductLineRepositoryImpl implements ProductLineRepository {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public ProductLine getProductLineById(String id) {
-        ProductLine productLine = entityManager.find(ProductLine.class, id);
-        return productLine;
+        return entityManager.find(ProductLine.class, id);
     }
 
     @Override
     @Transactional
-    public void saveProductLine(ProductLine productLine) {
+    public ProductLine saveProductLine(ProductLine productLine) {
         entityManager.persist(productLine);
+        return productLine;
     }
 
     @Override
     @Transactional
     public void deleteProductLine(String id) {
         ProductLine productLine = entityManager.find(ProductLine.class, id);
-        entityManager.remove(productLine);
+        if(productLine != null) {
+            entityManager.remove(productLine);
+        }
     }
 
     @Override
     @Transactional
-    public void updateProductLine(ProductLine productLine) {
-        entityManager.merge(productLine);
+    public ProductLine updateProductLine(ProductLine productLine) {
+        return entityManager.merge(productLine);
+    }
+
+    @Override
+    public List<ProductLine> getAllProductLines() {
+        List<ProductLine> productLines = new ArrayList<>();
+        TypedQuery<ProductLine> query = entityManager.createQuery("SELECT pl FROM ProductLine pl", ProductLine.class);
+        productLines = query.getResultList();
+        return productLines;
+    }
+
+    @Override
+    public List<ProductLine> getProductLinesByDescription(String description) {
+        List<ProductLine> productLines = new ArrayList<>();
+        TypedQuery<ProductLine> query = entityManager.createQuery("SELECT pl FROM ProductLine pl WHERE pl.textDescription LIKE :keyword", ProductLine.class);
+        query.setParameter("keyword", "%" + description + "%");
+        productLines = query.getResultList();
+        return productLines;
+    }
+
+    @Override
+    public List<ProductLine> getProductLinesWithImages() {
+        return null;
     }
 }
